Implemented MultiStartGenerator
