+++ bqskit/ir/circuit.py
@@ -28,9 +28,13 @@ from bqskit.ir.location import CircuitLocation
 from bqskit.ir.location import CircuitLocationLike
 from bqskit.ir.operation import Operation
 from bqskit.ir.opt.cost.functions import HilbertSchmidtCost
+from bqskit.ir.opt.cost.functions import HilbertSchmidtCostGenerator
+from bqskit.ir.opt.cost.generator import CostFunctionGenerator
 from bqskit.ir.opt.instantiater import Instantiater
 from bqskit.ir.opt.instantiaters import instantiater_order
 from bqskit.ir.opt.minimizers.ceres import CeresMinimizer
+from bqskit.ir.opt.multistartgen import MultiStartGenerator
+from bqskit.ir.opt.multistartgens.random import RandomStartGenerator
 from bqskit.ir.point import CircuitPoint
 from bqskit.ir.point import CircuitPointLike
 from bqskit.ir.region import CircuitRegion
@@ -2209,12 +2213,11 @@ class Circuit(DifferentiableUnitary, StateVectorMap, Collection[Operation]):
     def instantiate(
         self,
         target: StateLike | UnitaryLike,
-        method: str | None = None,
+        method: str | Instantiater | None = None,
         multistarts: int = 1,
         seed: int | None = None,
-        starts: list[npt.NDArray[np.float64]] | None = None,
-        cost_fn_gen: CostFunctionGenerator | None = None,
-        instantiater: Instantiater | None = None,
+        start_gen: MultiStartGenerator = RandomStartGenerator(),
+        cost_fn_gen: CostFunctionGenerator = HilbertSchmidtCostGenerator(),
         **kwargs: Any,
     ) -> Circuit:
         """
@@ -2261,81 +2264,81 @@ class Circuit(DifferentiableUnitary, StateVectorMap, Collection[Operation]):
 
             ValueError: If `seed` is not an integer or `None`
         """
-        # Set seed if specified
-        if seed is not None:
-            if not isinstance(seed, int):
-                raise ValueError(
-                    f'Expected seed to be an integer got {type(seed)}.',
-                )
-            seed_random_sources(seed)
+        # # Set seed if specified
+        # if seed is not None:
+        #     if not isinstance(seed, int):
+        #         raise ValueError(
+        #             f'Expected seed to be an integer got {type(seed)}.',
+        #         )
+        #     seed_random_sources(seed)
 
-        # Construct instantiater from method if it was not passed as an argument
-        err = ''
-        if instantiater is None:
-            for inst in instantiater_order:
-                # If method is specified; match it
-                if inst.get_method_name().lower() == method:  # type: ignore
-                    if not inst.is_capable(self):  # type: ignore
-                        raise ValueError(
-                            'Circuit cannot be instantiated using the '
-                            f'{method} method.'
-                            f'\n{inst.get_violation_report(self)}',  # type: ignore
-                        )
-                    instantiater = inst(**kwargs)
-                    break
+        # # Construct instantiater from method if it was not passed as an argument
+        # err = ''
+        # if instantiater is None:
+        #     for inst in instantiater_order:
+        #         # If method is specified; match it
+        #         if inst.get_method_name().lower() == method:  # type: ignore
+        #             if not inst.is_capable(self):  # type: ignore
+        #                 raise ValueError(
+        #                     'Circuit cannot be instantiated using the '
+        #                     f'{method} method.'
+        #                     f'\n{inst.get_violation_report(self)}',  # type: ignore
+        #                 )
+        #             instantiater = inst(**kwargs)
+        #             break
 
-                # If method is not specified; find first capable one
-                if method is None:
-                    if inst.is_capable(self):  # type: ignore
-                        instantiater = inst(**kwargs)
-                        break
-                    err += inst.get_violation_report(self) + '\n'  # type: ignore
+        #         # If method is not specified; find first capable one
+        #         if method is None:
+        #             if inst.is_capable(self):  # type: ignore
+        #                 instantiater = inst(**kwargs)
+        #                 break
+        # err += inst.get_violation_report(self) + '\n'  # type: ignore
 
-        if instantiater is None:
-            if err != '':
-                raise ValueError(f'No capable instantiater.\n{err}')
-            raise ValueError(f'No such instantiatation method {method}.')
+        # if instantiater is None:
+        #     if err != '':
+        #         raise ValueError(f'No capable instantiater.\n{err}')
+        #     raise ValueError(f'No such instantiatation method {method}.')
 
-        # Check Target
-        if is_square_matrix(target):
-            target = UnitaryMatrix(target)  # type: ignore
-        elif is_vector(target):
-            target = StateVector(target)  # type: ignore
-        else:
-            raise TypeError(
-                'Expected either StateVector or UnitaryMatrix'
-                ' for target, got %s.' % type(target),
-            )
+        # # Check Target
+        # if is_square_matrix(target):
+        #     target = UnitaryMatrix(target)  # type: ignore
+        # elif is_vector(target):
+        #     target = StateVector(target)  # type: ignore
+        # else:
+        #     raise TypeError(
+        #         'Expected either StateVector or UnitaryMatrix'
+        #         ' for target, got %s.' % type(target),
+        #     )
 
-        if target.dim != self.dim:
-            raise ValueError('Target dimension mismatch with circuit.')
+        # if target.dim != self.dim:
+        #     raise ValueError('Target dimension mismatch with circuit.')
 
-        # Generate starting points
-        if starts is None:
-            starts = instantiater.gen_starting_points(multistarts, self, target)
-        elif len(starts) != multistarts:
-            raise ValueError(f'Was told to expect {multistarts} starts but got {len(starts)}.')
+        # # Generate starting points
+        # if starts is None:
+        #     starts = instantiater.gen_starting_points(multistarts, self, target)
+        # elif len(starts) != multistarts:
+        #     raise ValueError(f'Was told to expect {multistarts} starts but got {len(starts)}.')
 
-        # Instantiate the circuit
-        params_list = []
-        for start in starts:
-            params_list.append(instantiater.instantiate(self, target, start))
+        # # Instantiate the circuit
+        # params_list = []
+        # for start in starts:
+        #     params_list.append(instantiater.instantiate(self, target, start))
 
-        # Return best result
-        if multistarts == 1:
-            self.set_params(params_list[0])
-            return self
+        # # Return best result
+        # if multistarts == 1:
+        #     self.set_params(params_list[0])
+        #     return self
 
-        if cost_fn_gen is not None:
-            if not isinstance(cost_fn_gen, CostFunctionGenerator):
-                raise TypeError(
-                        'Expected CostFunctionGenerator, got %s' % type(cost_fn_gen)
-                        )
-            cost_fn = cost_fn_gen.gen_cost(self, target)
-        else:
-            cost_fn = HilbertSchmidtCost(self, target)
+        # if cost_fn_gen is not None:
+        #     if not isinstance(cost_fn_gen, CostFunctionGenerator):
+        #         raise TypeError(
+        #                 'Expected CostFunctionGenerator, got %s' % type(cost_fn_gen)
+        #                 )
+        #     cost_fn = cost_fn_gen.gen_cost(self, target)
+        # else:
+        #     cost_fn = HilbertSchmidtCost(self, target)
 
-        self.set_params(sorted(params_list, key=lambda x: cost_fn(x))[0])
+        # self.set_params(sorted(params_list, key=lambda x: cost_fn(x))[0])
         return self
 
     def minimize(self, cost: CostFunction, **kwargs: Any) -> None:
