+++ bqskit/ir/opt/__init__.py
@@ -13,6 +13,8 @@ from bqskit.ir.opt.instantiaters import Minimization
 from bqskit.ir.opt.instantiaters import QFactor
 from bqskit.ir.opt.minimizer import Minimizer
 from bqskit.ir.opt.minimizers import LBFGSMinimizer
+from bqskit.ir.opt.multistartgen import MultiStartGenerator
+from bqskit.ir.opt.multistartgens import RandomStartGenerator
 
 __all__ = [
     'Instantiater',
@@ -26,4 +28,6 @@ __all__ = [
     'HilbertSchmidtCostGenerator',
     'HilbertSchmidtResiduals',
     'HilbertSchmidtResidualsGenerator',
+    'MultiStartGenerator',
+    'RandomStartGenerator',
 ]
