+++ bqskit/ir/opt/multistartgens/__init__.py
@@ -1,0 +1,6 @@
+"""This package implements multistart generators."""
+from __future__ import annotations
+
+from bqskit.ir.opt.multistartgens.random import RandomStartGenerator
+
+__all__ = ['RandomStartGenerator']
